requirements.txt
langchain==0.1.1
langchain-core==0.1.12
langchain-google-genai==0.0.6
pydantic==2.5.2
pyyaml==6.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
asyncio>=3.4.3
requests>=2.31.0
langchain-community>=0.0.10
simpleeval